Fixed ratio plot, added movie and export AVI
ratio plot had a fraction that was inverted, giving nonsensical plots.
Added a new function that will ListAnimate a movie and export it as an
AVI. This seemed to be a good idea so the movie could be view and
exported without having to produce the movie frames two separate times.