Added orientation flipping to movement
When moving the HTMF or the cinnamate, it will now see if a possible
orientation flip is favorable at the same time as checking the move. It
still won’t flip sign in place, but at least this will allow the
product distribution to not be limited by the pseudorandom starting
conditions.